Updates command-line argument parser to produce more useful values.
Improves test for command line arguments.

This also includes a test fixture that creates temporary files which can be used as a model in the future.  If time becomes a concern, remember that pytest has a session-scoped factory method that is meant to be extended by other tests and fixtures.

For some reason, the type checker is insisting that _pathlib.Path_ objects are not os.PathLike so this also includes comments to disable type checking on the apporpriate lines.