Creates CSV Globbing functionality (#19)
The primary purpose of this commit is to include a function that robustly retrieves all relevant CSV files from a provided directory path.  Searches can be recursive, and the function accepts a file to exclude from the results.

This pull request also normalizes the results of the command-line argument parser so that all the appropriate arguments are converted into _pathlib.Path_ objects.

This improves testing, including tests for the new functionality.  It also includes some directives to smooth type checking because, for reasons unknown, _Path_ objects are not _os.PathLike_.