CSV Filter, Merge & Archive (#23)
This breaks apart the basic functionality of the program to make it more modular and easier to test and provides basic tests for it.

The nature of the application file-system intensive, so testing can be a challenge. Some contortions were required to separate file system operations from business logic so that they could be tested independently. Some test fixtures were created to provide a common file structure and reduce the burden of adding more tests.

This also includes several fixes for bugs that were revealed by the added tests.